backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
distro==1.9.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Limiter==4.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
limits==5.6.0
MarkupSafe==3.0.3
openai==2.16.0
ordered-set==4.1.0
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
python-dotenv==1.2.1
sniffio==1.3.1
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.5
wrapt==2.1.0

# server
gunicorn==21.2.0
